requirements.txt
customtkinter
openpyxl
reportlab
PyPDF2
pymupdf
Pillow